+++ packages/eslint-plugin-calcite-components/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.0-next.3](https://github.com/Esri/calcite-design-system/compare/@esri/eslint-plugin-calcite-components@2.0.0-next.2...@esri/eslint-plugin-calcite-components@2.0.0-next.3) (2025-01-01)
+
+**Note:** Version bump only for package @esri/eslint-plugin-calcite-components
+
+## [2.0.0-next.2](https://github.com/Esri/calcite-design-system/compare/@esri/eslint-plugin-calcite-components@2.0.0-next.1...@esri/eslint-plugin-calcite-components@2.0.0-next.2) (2024-12-28)
+
+**Note:** Version bump only for package @esri/eslint-plugin-calcite-components
+
 ## [2.0.0-next.1](https://github.com/Esri/calcite-design-system/compare/@esri/eslint-plugin-calcite-components@2.0.0-next.0...@esri/eslint-plugin-calcite-components@2.0.0-next.1) (2024-11-15)
 
 ### ⚠ BREAKING CHANGES